+++ Assets/Scripts/AbilitySystem/Modifiers/ExplosiveShotAbility.cs
@@ -5,7 +5,7 @@ public class ExplosiveShotAbility : Ability
 {
     public GameObject explosionPrefab;
     public float explosionRadius = 5f;
-    public int explosionDamage = 50;
+    public float explosionDamage = 50f;
     public LayerMask damageLayers;
     [Tooltip("Orient the explosion to face back toward the direction the projectile came from.")]
     public bool orientToImpact = false;
