+++ Assets/Scripts/AbilitySystem/Modifiers/ShieldExplosionAbility.cs
@@ -5,7 +5,7 @@ public class ShieldExplosionAbility : Ability
 {
     public GameObject explosionPrefab;
     public float explosionRadius = 6f;
-    public int explosionDamage = 50;
+    public float explosionDamage = 50f;
     public LayerMask damageLayers;
     public bool explodeOnBreak = true;
     public bool explodeOnExpire = false;
